app.py
# Import necessary libraries
import streamlit as st
import numpy as np
from PIL import Image

# Import functions from zoom.py
from zoom import process_image, yolo, face_complete_distance_center_light, glass_detection

# If using a ML model, import TensorFlow, PyTorch, etc.
# from tensorflow.keras.models import load_model

# Set up the logo, title and introduction of the app
st.image("images/logo.png", width=150,)
st.title('First Impression for Videocalls')
st.write('This application provides insights into your first impression in videocalls based on the uploaded image.')

# Image Upload Section
uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "png", "jpeg"])
if uploaded_file is not None:
    # Display the uploaded image
    image = Image.open(uploaded_file)
    st.image(image, caption='Your Uploaded Image', use_column_width=True)

    # Convert the uploaded file to a format that can be processed by zoom.py
    # For example, save the image to a temporary file
    with open("temp_image.jpg", "wb") as f:
        f.write(uploaded_file.getbuffer())

    # Call the process_image function from zoom.py
    process_image("temp_image.jpg")

    # Display results - Modify this part based on how you want to present the results
    st.subheader('Analysis Results')
    st.write('Based on our analysis...')
    # You might want to modify process_image in zoom.py to return results instead of printing them
else:
    st.warning('Please upload an image.')

# Additional Guidelines (if needed)
st.subheader('Guidelines')
st.image("images/background.png", width=350)
st.write('Here are some guidelines for selecting appropriate images for analysis...')
# Add more instructions or use st.image, st.video for additional resources.
